refactor: Rename Calendar component appropriately
- Keep in a separate commit to preserve its git history
